+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 - Close public access to TarantoolResult*Impl ([#326](https://github.com/tarantool/cartridge-java/issues/326))
+- Add deep copy instead of shallow copy in default message pack mapper
+
 ## [0.10.1] - 2023-01-13
 
 ### Internal and API changes